fix exmem build for r1153
git-svn-id: http://tools.assembla.com/svn/chdk/trunk@1156 6794e30b-3f2a-0410-a806-a2bbca1c07ff

+++ platform/generic/wrappers.c
@@ -723,7 +723,7 @@ void exmem_malloc_init() {
 	}
 }
 
-void *malloc(unsigned size) {
+void *malloc(long size) {
 	if(exmem_heap)
 		return suba_alloc(exmem_heap,size,0);
 	else
